Move compute_scaling_rates to shared
